% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_AUC.R
\name{Calc_AUC}
\alias{Calc_AUC}
\title{function area under the curve}
\usage{
Calc_AUC(Dat_Frm, Key.Name, Key.List, Foc.Var, Out.Var, date.min = NULL,
  date.max = NULL)
}
\description{
function area under the curve
}
