% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est.Tise.loess.R
\name{Est.Tise.loess}
\alias{Est.Tise.loess}
\title{function estimated time series using loess method 071917}
\usage{
Est.Tise.loess(Dat_Frm, Key.Name, Key.List, Foc.Var, date.min = NULL,
  date.max = NULL, loess.span = NULL)
}
\description{
function estimated time series using loess method 071917
}
