% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTise.R
\name{EstTise}
\alias{EstTise}
\title{wrapper around Yval.sps
use Yval.sps as internal}
\usage{
EstTise(Data_Frame, sp_list, Var_Interest = NULL, method = NULL)
}
\description{
wrapper around Yval.sps
use Yval.sps as internal
}
