% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrpl.R
\name{intrpl}
\alias{intrpl}
\title{interpolation function __intrpl__
function adding itplation of IntrPol_var to data fame, with two keys _Prime_key and Secnd_Key_ from Time_full range, find missing
time points and fill data from missing time points, uses date time variable}
\usage{
intrpl(Dat_Frm, Data_Type, IntrPol_Var, Time_Var = NULL, Prime_Key = NULL,
  Secnd_Key = NULL, Time_Full = NULL)
}
\description{
interpolation function __intrpl__
function adding itplation of IntrPol_var to data fame, with two keys _Prime_key and Secnd_Key_ from Time_full range, find missing
time points and fill data from missing time points, uses date time variable
}
