% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.plot.gen.R
\name{regress.plot.gen}
\alias{regress.plot.gen}
\title{============================================================================================================
function ..regress.plot.gen()
 this function is to generate "list" contain multiple plots with single x axis and multlie y varibes.
 output is "list"
 output can be recalled within multiplot function to generate figure
 for example, f<-regress.plot.gen(df,x_name,y_list); mutliplot(plotlist=f, cols=N) will give desirable plots
============================================================================================================}
\usage{
regress.plot.gen(df_name, x_name, y_list, xlabname = NULL)
}
\description{
============================================================================================================
function ..regress.plot.gen()
 this function is to generate "list" contain multiple plots with single x axis and multlie y varibes.
 output is "list"
 output can be recalled within multiplot function to generate figure
 for example, f<-regress.plot.gen(df,x_name,y_list); mutliplot(plotlist=f, cols=N) will give desirable plots
============================================================================================================
}
