% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setValueLimit.R
\name{setValueLimit}
\alias{setValueLimit}
\title{Function for correcting values out of range}
\usage{
setValueLimit(array, lowerlimit, upperlimit)
}
\description{
Function for correcting values out of range
}
