% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est.Tise.loess.R
\name{Est.Tise.loess}
\alias{Est.Tise.loess}
\title{function estimated time series using loess method 071917
Dat_Frm: data frame of interest
Key.Name: Key
Key.List: List of Key variables to pass
Foc.Var: Variable of interest to get Area Under the Curve
Out.Var: Variable}
\usage{
Est.Tise.loess(Dat_Frm, Key.Name, Key.List, Foc.Var, date.min = NULL,
  date.max = NULL, ...)
}
\description{
function estimated time series using loess method 071917
Dat_Frm: data frame of interest
Key.Name: Key
Key.List: List of Key variables to pass
Foc.Var: Variable of interest to get Area Under the Curve
Out.Var: Variable
}
